chore-서버 구조 변경
